Fixed #32: reseting logs every hour
metrics image is not easily configurable, with several services running in parallel
and none explicitly writting to /logs/all.log which is growing forever.
It seems to be generated by some python code but what code triggering it and if/where
to configure it is unknown.

Instead, a simple logs reset mechanism has been added to metrics image which can be
enabled using `TRUNCATE_LOGS` environment variable.