Updated `get_zim_files` iterator to never return hidden ZIM files
Code has a `with_hidden` param that allows returning hidden (dot-prefixed) folders
and files.
This was added to support the .hidden folder for dev-library.

It was actually never intended (I believe!) that we'd support hidden ZIM files.
It serves no existing use case.

This updates allows creating hidden .zim file in the root tree yet not having
them included in the catalog.