Updated bittorrent tracker to torrust
Changed tracker software from opentracker to more modern torrust

Using an upstream image strategy with a side-container doing the whitelist management
Whitelister is a python script that runs forever, periodically queries OPDS
and calls the tracker's API endpoint to add each entry to the whitelist.

There's also the complete setup of torrust-gui for reference as it is not
what we want.
This is pushed as we've not discussed the future of this feature.