swap library files using rename to avoid empty file edge case
When we generate libraries (public and internal ones), we write to a temp file first,
ensures the markup is OK and then replace final path with tmp one.

We used to have the temp files within TMPDIR and thus replace by copying as this would
likely cross filesystems (final files are in mounted volumes)

This has the unfortunate consequence of having the file in incomplete state during the
short duration of the write on disk.

We are thus now using predictable, next-to-final temp files so those can be renamed
instead of copied, avoiding this issue