When running for ALL, only consider ZIMs that are ready on mirrorbrain
As we rely more and more on bittorrent for ZIM distribution, it is
important that any ZIM in the Kiwix Catalog can be downloaded by BT.
Given the lib generation is filesystem based, we could see new ZIM files
on the filesystem before mirrorbrain had a chance to record it in its
DB (rare) or could compute all hashes that it needs for it to be avail
via BT (more frequent).

With this change, when listing ZIM files on the fs, we query
the mirrorbrain database (PG) to check if it's BT ready (has a
corresponding entry in hashes table).

This results in ~40 SELECT queries to mirrorbrain that are sent
in a single transaction.

It's not invasive as it happens after the filesystem read step and
simply filters out entries not in MB