Fixed library generator adding incorrect book entries to catalog
Previous impl. of restricting entries to the mirrorbrain-ready ones
was done post readfs step in order to be non-invasive:
- reads filesystem
- then query MB and pop non-mb-ready entries out
- then write catalog

Issue was that readfs() only reads the ZIM metadata
from the first (ie. most recent) ZIM for an alias (because its slow)
but any non-mb-ready entry would be the most recent for its alias and
would thus have only been fs-read but not ZIM-read.

This lead to producing catalog with entries like
<book size="xxx" url="xxx" faviconMimeType="image/png"/>
for all entries matching an alias for which there was a non-mb-ready ZIM

This has been fixed by applying the exclusion before reading the ZIMs,
just after iterating the list of files.

There's also a safeguard check when writing catalog to not do it
should an entry lack basic metadata such as ID