Use HTTPS only on mirrorbrain
- We have been supporting both HTTP and HTTPs for a long time.
- In 065f8e71 we made were not degrading protocol from HTTPS to HTTP
- All our mirrors are HTTPs so anything not served directly by MB
is not usable without HTTPs
- We switched all other services long ago

MB being an HTTP-only apache behind an SSL-enabled reverse proxy
thought it was HTTP only and thus, in addition to redirects (mentioned
above and handled), it was only generating HTTP links.

That's the case for instance for webseed URL in magnet.

The consequence is that whenever using an endpoint served by MB which
generates links (metalink, magnet for instance), user would downgrade
and thus exposes himself before going back to HTTPs (mirrors).

Additionally, this had the unfortunate consequence of kiwix-serve on
library.kiwix.org to not be able to parse metalink (JS's `fetch()`
is not allowed to downgrade) and thus always fell back on the simpler
magnet generation which only contains the MB link as webseed instead
of the list of mirrored-links

This introduce two change:
- MB now returns HTTPs links everywhere.
- download.kiwix.org now redirects HTTP to HTTPS