Allowing browse clients to read /skin/ without anubis
/skin/ can be read very early, before the client is cleared by anubis
and this could result in some resources not being fetched properly.
I've seen favicons, manifest and even CSS dependency resulting
in broken UI that required a refresh.

This change sets the anubis version we use. Unfortunately, they
dont tag images so it's not straightforward to know which version we
use. This is release 1.22.0.

As we need a specific rule, we can't use the default config and thus
have to bring a copy of the config. This will need to be maintained
as well…

A single new rule allowing HEAD+GET on ^/skin/ is now in place

Fixes https://github.com/kiwix/operations/issues/470